refactor: move Simd.h, BatchAccessor.h, StencilAccessor.h to nanovdb/util/
All three headers are now shared dependencies across multiple examples and
are no longer specific to a single example directory.  Consolidate into
nanovdb/nanovdb/util/ to match the existing NanoVDB utility layout.

- simd_test/Simd.h → nanovdb/util/Simd.h
- ex_voxelBlockManager_host_cuda/BatchAccessor.h → nanovdb/util/BatchAccessor.h
- ex_voxelBlockManager_host_cuda/StencilAccessor.h → nanovdb/util/StencilAccessor.h

Update all #include paths in BatchAccessor.h, StencilAccessor.h,
stencil_gather_cpu.cpp, and simd_test scratchpad files.

Also add HaloStencilAccessor.md: design document for the dense halo buffer
approach to CPU WENO5 stencil extraction (successor to StencilAccessor).
